Fix C++ abort on bad magic, thread Java JDK17 toolchain, correct test counts
- src/cpp/main.cpp: wrap post-parse dispatch in try/catch(DeltaError) so
  malformed magic exits 1 with a message instead of aborting (Abort trap: 6)
- tests/correctness.sh: detect JDK 17 javac and pass JAVA=/JAVAC= to
  make test; bare make test was resolving to system javac which rejects
  --release 17; also update all test counts to actuals (236/89/93/230/73/81)
- tests/per-language-benchmark.sh: derive JAVAC from same JDK as JAVA
  and pass both to make -s when building the Java implementation
- ANALYSIS.md: update test scripts table to 236/89/93/230/73/81
- .gitignore: exclude *-REVIEW.md and RESPONSE.md
- Remove Kotlin and Scala source trees (src/kotlin/, src/scala/) and
  all references in scripts and docs; fix Rust/Go test drift against
  the 6-arg encode_delta_large API; fix Java long truncation bug and
  stale-bytecode detection in test_delta.sh

+++ src/java/delta/Diff.java
@@ -28,7 +28,7 @@ public final class Diff {
 
     /** Print delta compression statistics to stderr. */
     static void printStats(List<Command> commands) {
-        List<Integer> copyLens = new ArrayList<>();
+        List<Long> copyLens = new ArrayList<>();
         long totalCopy = 0, totalAdd = 0;
         int numCopies = 0, numAdds = 0;
         for (Command cmd : commands) {
@@ -49,7 +49,7 @@ public final class Diff {
         if (!copyLens.isEmpty()) {
             copyLens.sort(null);
             double mean = totalCopy / (double) copyLens.size();
-            int median = copyLens.get(copyLens.size() / 2);
+            long median = copyLens.get(copyLens.size() / 2);
             System.err.printf("  copies: %d regions, min=%d max=%d mean=%.1f median=%d bytes%n",
                 copyLens.size(), copyLens.get(0), copyLens.get(copyLens.size() - 1), mean, median);
         }
